Django/requirements.txt
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bootstrap4==0.1.0
Brotli==1.0.7
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-appconf==1.0.3
django-extensions==2.2.3
django-imagekit==4.0.2
django-mysql==3.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
et-xmlfile==1.0.1
gunicorn==19.9.0
heroku==0.1.4
idna==2.8
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
jdcal==1.4
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
lazy-object-proxy==1.4.1
lxml==4.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
oauthlib==3.1.0
openpyxl==2.6.2
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
pilkit==2.0
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pyperclip==1.7.0
pyrsistent==0.15.3
python-dateutil==1.5
python3-openid==3.1.0
pytz==2019.3
pywinpty==0.5.5
pyzmq==18.0.2
qtconsole==4.5.2
requests==2.22.0
requests-oauthlib==1.2.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.0
wrapt==1.11.2
